Refactor services into packages and add modular OCR support
